requirements.txt
pandas==1.4.1
scikit-learn==1.0.2
python-Levenshtein==0.12.2
